level 1- *added SplashScreen to prevent AuthActivity screen from being shown if the user has been saved
