+++ app/src/main/java/com/project/myapp/screens/main/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import com.project.myapp.KeysHolder.USER_NAME_KEY
 import com.project.myapp.databinding.ActivityMainBinding
-import com.project.myapp.ext.componentactivity.EnableEdgeToEdgeGrayStatusBar
+import com.project.myapp.ext.componentactivity.enableEdgeToEdgeGrayStatusBar
 import com.project.myapp.ext.componentactivity.handleBackPress
 import com.project.myapp.ext.context.customAnimationBackward
 import com.project.myapp.ext.view.initializeWindowInsetsHandling
@@ -31,7 +31,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun setView() {
-        EnableEdgeToEdgeGrayStatusBar()
+        enableEdgeToEdgeGrayStatusBar()
         setContentView(binding.root)
         binding.root.initializeWindowInsetsHandling()
     }
