+++ app/src/main/res/drawable/splash_screen.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@color/primary_background" />
+<!--    <item>-->
+<!--        <bitmap-->
+<!--            android:gravity="center"-->
+<!--            android:src = "@drawable/ic_launcher_round" />-->
+<!--    </item>-->
+</layer-list>
