+++ app/src/main/res/values/themes.xml
@@ -5,4 +5,8 @@
     </style>
     <style name="Theme.MyApp" parent="Base.Theme.MyApp" />
 
+    <style name="SplashTheme" parent="Theme.MyApp">
+        <item name="android:windowBackground">@drawable/splash_screen</item>
+    </style>
+
 </resources>
